requirements.txt
pypdf
gradio
google-genai
python-dotenv